reorganized the plan entity UI to move common info into the top (always visible) section
