+++ manager/ui/war/src/main/java/io/apiman/manager/ui/client/local/pages/AbstractPlanPage.java
@@ -15,8 +15,11 @@
  */
 package io.apiman.manager.ui.client.local.pages;
 
+import io.apiman.manager.api.beans.actions.ActionBean;
+import io.apiman.manager.api.beans.actions.ActionType;
 import io.apiman.manager.api.beans.idm.PermissionType;
 import io.apiman.manager.api.beans.plans.PlanBean;
+import io.apiman.manager.api.beans.plans.PlanStatus;
 import io.apiman.manager.api.beans.plans.PlanVersionBean;
 import io.apiman.manager.api.beans.summary.PlanVersionSummaryBean;
 import io.apiman.manager.ui.client.local.AppMessages;
@@ -24,6 +27,7 @@ import io.apiman.manager.ui.client.local.pages.common.Breadcrumb;
 import io.apiman.manager.ui.client.local.pages.common.VersionSelector;
 import io.apiman.manager.ui.client.local.services.ContextKeys;
 import io.apiman.manager.ui.client.local.services.rest.IRestInvokerCallback;
+import io.apiman.manager.ui.client.local.util.Formatting;
 import io.apiman.manager.ui.client.local.util.MultimapUtil;
 
 import java.util.ArrayList;
@@ -34,10 +38,15 @@ import javax.inject.Inject;
 
 import org.jboss.errai.ui.nav.client.local.PageState;
 import org.jboss.errai.ui.shared.api.annotations.DataField;
+import org.jboss.errai.ui.shared.api.annotations.EventHandler;
+import org.overlord.commons.gwt.client.local.widgets.AsyncActionButton;
 
+import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.event.logical.shared.ValueChangeEvent;
 import com.google.gwt.event.logical.shared.ValueChangeHandler;
 import com.google.gwt.user.client.ui.Anchor;
+import com.google.gwt.user.client.ui.InlineLabel;
+import com.google.gwt.user.client.ui.Label;
 
 
 /**
@@ -68,6 +77,20 @@ public abstract class AbstractPlanPage extends AbstractPage {
     @Inject @DataField
     Anchor toNewPlanVersion;
     
+    @Inject @DataField
+    Label description;
+    @Inject @DataField
+    InlineLabel createdOn;
+    @Inject @DataField
+    Anchor createdBy;
+    @Inject @DataField
+    InlineLabel status;
+
+    @Inject @DataField
+    Anchor ttd_toNewVersion;
+    @Inject @DataField
+    AsyncActionButton lockButton;
+
     @Inject @DataField
     Anchor toPlanOverview;
     @Inject @DataField
@@ -197,8 +220,107 @@ public abstract class AbstractPlanPage extends AbstractPage {
         breadcrumb.addItem(dashHref, "home", i18n.format(AppMessages.HOME)); //$NON-NLS-1$
         breadcrumb.addItem(orgPlansHref, "shield", versionBean.getPlan().getOrganization().getName()); //$NON-NLS-1$
         breadcrumb.addActiveItem("bar-chart-o", planBean.getName()); //$NON-NLS-1$
+
+        String newVersionHref = navHelper.createHrefToPage(NewPlanVersionPage.class, MultimapUtil.fromMultiple("org", org, "plan", plan)); //$NON-NLS-1$ //$NON-NLS-2$
+        ttd_toNewVersion.setHref(newVersionHref);
+
+        description.setText(planBean.getDescription());
+        createdOn.setText(Formatting.formatShortDate(versionBean.getCreatedOn()));
+        createdBy.setText(versionBean.getCreatedBy());
+        String toUserHref = navHelper.createHrefToPage(UserServicesPage.class,
+                MultimapUtil.fromMultiple("user", planBean.getCreatedBy())); //$NON-NLS-1$
+        createdBy.setHref(toUserHref);
     }
 
+    /**
+     * @see io.apiman.manager.ui.client.local.pages.AbstractPage#onPageLoaded()
+     */
+    @Override
+    protected void onPageLoaded() {
+        super.onPageLoaded();
+        lockButton.reset();
+        renderPlanStatus();
+    }
+
+    /**
+     * Updates various UI bits based on the status of the plan.
+     */
+    protected void renderPlanStatus() {
+        status.setText(versionBean.getStatus().toString());
+        status.setTitle(getStatusDescription(versionBean.getStatus()));
+        setStatusLabelClass(status, versionBean.getStatus());
+
+        boolean canLock = versionBean.getStatus() == PlanStatus.Ready || versionBean.getStatus() == PlanStatus.Created;
+        boolean locked = versionBean.getStatus() == PlanStatus.Locked;
+        lockButton.setEnabled(canLock);
+        lockButton.setVisible(!locked);
+    }
+
+    /**
+     * Called when the user clicks the "Lock" button.
+     * @param event
+     */
+    @EventHandler("lockButton")
+    public void onLock(ClickEvent event) {
+        lockButton.onActionStarted();
+
+        ActionBean action = new ActionBean();
+        action.setType(ActionType.lockPlan);
+        action.setOrganizationId(versionBean.getPlan().getOrganization().getId());
+        action.setEntityId(versionBean.getPlan().getId());
+        action.setEntityVersion(versionBean.getVersion());
+        rest.performAction(action, new IRestInvokerCallback<Void>() {
+            @Override
+            public void onSuccess(Void response) {
+                lockButton.onActionComplete();
+                versionBean.setStatus(PlanStatus.Locked);
+                status.setText(PlanStatus.Locked.toString());
+                renderPlanStatus();
+                hideElementsBasedOnStatus();
+            }
+            
+            @Override
+            public void onError(Throwable error) {
+                dataPacketError(error);
+            }
+        });
+    }
+
+    /**
+     * Sets the proper CSS class(es) on the label based on the plan's status.
+     * @param label
+     * @param status
+     */
+    private static void setStatusLabelClass(InlineLabel label, PlanStatus status) {
+        label.getElement().setClassName("apiman-label"); //$NON-NLS-1$
+        switch (status) {
+        case Created:
+        case Ready:
+            label.getElement().addClassName("apiman-label-warning"); //$NON-NLS-1$
+            break;
+        case Locked:
+            label.getElement().addClassName("apiman-label-success"); //$NON-NLS-1$
+            break;
+        }
+    }
+
+    /**
+     * Gets an explanation of the status.
+     * @param status
+     */
+    private String getStatusDescription(PlanStatus status) {
+        switch (status) {
+        case Created:
+            return i18n.format(AppMessages.PLAN_STATUS_CREATED);
+        case Ready:
+            return i18n.format(AppMessages.PLAN_STATUS_CREATED);
+        case Locked:
+            return i18n.format(AppMessages.PLAN_STATUS_LOCKED);
+        default:
+            return null;
+        }
+    }
+    
     /**
      * @return a list of versions
      */
